% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dynamic_title_doughnut}
\alias{get_dynamic_title_doughnut}
\title{Adjust voting title properties based on text length}
\usage{
get_dynamic_title_doughnut(title_text)
}
\arguments{
\item{title_text}{The character string for the title.}
}
\value{
A list containing the wrapped text (\code{text}), the calculated font size (\code{size}), and the vertical justification value (\code{vjust}).
}
\description{
A specialized version of get_dynamic_title tailored to the larger font sizes
and different layout requirements of the voting graphs.
}
\examples{
get_dynamic_title_votes("Who is the most likely to lead a Martian colony?")
}
