% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dynamic_title_votes}
\alias{get_dynamic_title_votes}
\title{Version for voting title}
\usage{
get_dynamic_title_votes(title_text)
}
\arguments{
\item{title_text}{The raw string for the title.}
}
\value{
A list containing the wrapped text, calculated size, and vjust.
}
\description{
Version for voting title
}
