% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_color_light}
\alias{is_color_light}
\title{Check if a color is light}
\usage{
is_color_light(hex, threshold = 0.85)
}
\arguments{
\item{hex}{A character vector of hex color codes or R color names.}

\item{threshold}{A numeric value between 0 and 1. Colors with luminance above this value are considered light. Defaults to 0.85.}
}
\value{
A logical vector, \code{TRUE} for each color that is considered light.
}
\description{
Calculates the perceived luminance of a color and compares it to a threshold.
}
\examples{
is_color_light("#fefb41") # Bright yellow, should be TRUE
is_color_light("blue") # Dark color, should be FALSE
}
