% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tg_progress_done}
\alias{tg_progress_done}
\title{Finalize the progress bar output.}
\usage{
tg_progress_done(bar)
}
\arguments{
\item{bar}{Progress bar state from \code{tg_progress_bar}.}
}
\value{
NULL invisibly.
}
\description{
Finalize the progress bar output.
}
