% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tg_progress_update}
\alias{tg_progress_update}
\title{Update the progress bar state.}
\usage{
tg_progress_update(bar)
}
\arguments{
\item{bar}{Progress bar state from \code{tg_progress_bar}.}
}
\value{
Updated progress bar state.
}
\description{
Update the progress bar state.
}
