% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TG_votes.R
\name{TG_votes}
\alias{TG_votes}
\title{Create and Save a Circular Bar Chart for Votes/Opinion Data}
\usage{
TG_votes(
  dataset,
  column_name,
  title = NULL,
  name = "names",
  color = "favourite_color",
  sort_order = "desc",
  plot_zoom_mod = 0,
  name_size_mod = 0,
  title_size = NULL,
  title_vjust = NULL,
  output_path = "votes_plot.jpg",
  output_width = 7,
  output_height = 6,
  output_dpi = 300,
  save_plot = TRUE,
  show_plot = TRUE
)
}
\description{
Create and Save a Circular Bar Chart for Votes/Opinion Data
}
