% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraitGraph-package.R
\docType{package}
\name{TraitGraph-package}
\alias{TraitGraph}
\alias{TraitGraph-package}
\title{TraitGraph: Group Psychology Graph Maker}
\description{
Provides a suite of functions to generate publication-quality visualizations comparing group members on psychological traits or opinions. It allows for flexible customization of aesthetics and supports various graph layouts to highlight interpersonal dynamics and trait distributions.
}
\author{
\strong{Maintainer}: Lynden Jensen \email{lynden.jensen@gmail.com} (\href{https://orcid.org/0000-0003-0668-2543}{ORCID})

}
\keyword{internal}
