% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_color_light}
\alias{is_color_light}
\title{Check if a color is light}
\usage{
is_color_light(hex)
}
\arguments{
\item{hex}{A vector of hex color codes.}
}
\value{
A logical vector, TRUE if the color is considered light.
}
\description{
Check if a color is light
}
