% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TG_votes.R
\name{TG_votes}
\alias{TG_votes}
\title{Create and Save a Circular Bar Chart for Votes/Opinion Data}
\usage{
TG_votes(
  dataset,
  column_name,
  title = NULL,
  title_face = "bold",
  title_color = "black",
  name = "name",
  color = "favourite_color",
  sort_order = "desc",
  minor_ticks = 0,
  plot_zoom_mod = 0,
  name_size_mod = 0,
  name_position_mod = 1,
  title_size_mod = 0,
  title_vjust_mod = 0,
  output_path = "votes_plot.jpg",
  output_width = 7,
  output_height = 6,
  output_dpi = 300,
  save_plot = TRUE,
  show_plot = TRUE
)
}
\arguments{
\item{dataset}{A data frame containing the data to plot.}

\item{column_name}{The name of the column containing the numeric values to plot.}

\item{title}{A string for the plot's main title. If not provided, the name
of the \code{column_name} will be used.}

\item{title_face}{The font style of the title (e.g., \code{plain}, \code{bold}, or \code{bold.italic}). Defaalts to \code{bold}.}

\item{name}{The name of the column containing unique identifiers. Defaults to "names".}

\item{color}{The name of the column containing hex color codes. Defaults to "favourite_color".}

\item{sort_order}{The order to sort the results. Can be "desc" or "asc".}

\item{minor_ticks}{The number of minor y-axis ticks you want to display. Defaults to none.}

\item{plot_zoom_mod}{A numeric value to add or subtract from the default outer plot boundary.}

\item{name_size_mod}{A numeric value to add or subtract from the name label font size.}

\item{name_position_mod}{A numeric value to modify the vertical position of the individual names.}

\item{title_size_mod}{A numeric value to add or subtract from the dynamic title font size.}

\item{title_vjust_mod}{A numeric value to add or subtract from the dynamic title vertical adjustment.}

\item{output_path}{The full path where the plot will be saved.}

\item{output_width}{The width of the saved image in inches.}

\item{output_height}{The height of the saved image in inches.}

\item{output_dpi}{The resolution (dots per inch) for the saved image.}

\item{save_plot}{A logical value. If TRUE, the plot is saved to disk.}

\item{show_plot}{A logical value. If TRUE, the plot is displayed.}
}
\value{
Invisibly returns the ggplot object.
}
\description{
Create and Save a Circular Bar Chart for Votes/Opinion Data
}
